+++ packages/plugins/plugin-github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dependabit/plugin-github
 
+## 0.1.2
+
+### Patch Changes
+
+- Add gitignore-aware exclusions for manifest detection
+- Updated dependencies
+  - @dependabit/github-client@0.1.2
+  - @dependabit/manifest@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes