+++ packages/plugins/plugin-openapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dependabit/plugin-openapi
 
+## 0.1.2
+
+### Patch Changes
+
+- Add gitignore-aware exclusions for manifest detection
+- Updated dependencies
+  - @dependabit/manifest@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes