+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dependabit/test-utils
 
+## 0.1.2
+
+### Patch Changes
+
+- Add gitignore-aware exclusions for manifest detection
+
 ## 0.1.1
 
 ### Patch Changes